requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery>=1.17.0
google-cloud-storage>=1.17.0
